Add files via upload
This is who where my stolen intellectual property  copyright fraudulent fake ID THIEFT of Adrian C. Eugene...github is still under the law of Cftc as this is location is in American under Stephan j. Nix and KellyR.Hallman the cftc red list... I would love to use the pool and get paid that 120 grand that was in my paypal and keep everybody but them two. Turn them in for fraud... call me at 864-612-3976. Stephan joshua nix is a. NARCISSISTIC PERSON  BEWARE . HE LIVES HERE IN AMERICAN HE IS A LIAR.... HAVEA NICE DAY...$ACEINKICONS40 IS MY  CASH app...never got paid.would love to save my tattoo shop they are destroying with that money and make this project make a profit.